requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-docx==1.0.1
python-multipart==0.0.9
requests==2.31.0
pydantic
pytest
fastapi-cors==0.0.6